Add support for --statistics and --progressive options.
Now, if the window is too small, the spilled records are tacked onto
the end and message is written to stderr indicating how many records
were written out of sort order together with a hint about the size
of a --window parameter that could be used to avoid the failed sort. The
command exits with a non-zero exit code.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>